test(ui): split keyboard focus coverage
