+++ test/vitest/vitest.ui-isolated-paths.mjs
@@ -17,6 +17,7 @@ export const uiIsolatedTestFiles = [
   "ui/src/pages/chat/chat-pane-catalog.test.ts",
   "ui/src/pages/chat/chat-pane-history.test.ts",
   "ui/src/pages/chat/chat-pane-identity.test.ts",
+  "ui/src/pages/chat/chat-pane-keyboard-focus.test.ts",
   "ui/src/pages/chat/chat-pane-lifecycle.test.ts",
   "ui/src/pages/chat/chat-pane-pull-requests.test.ts",
   "ui/src/pages/chat/chat-pane-retained-presentation.test.ts",
