+++ ui/src/pages/chat/chat-pane.test.ts
@@ -8,7 +8,6 @@ import { createInitialUserMessageHandoff } from "../../app/initial-user-message-
 import { t } from "../../i18n/index.ts";
 import { showToast } from "../../lib/toast.ts";
 import {
-  getRenderedModalDialog,
   installDialogPolyfill,
   submitInputDialog,
   waitForConfirmDialogActions,
@@ -746,162 +745,6 @@ describe("chat pane initialization", () => {
 });
 
 describe("chat pane keyboard shortcuts", () => {
-  it("keeps the letter-to-composer contract when a button is focused", () => {
-    const { pane } = createTestChatPane({
-      client: createGatewayBrowserClientFixture(),
-      sessions: createSessionCapabilityFixture(),
-    });
-    pane.active = true;
-    pane.presented = true;
-    const composer = document.createElement("div");
-    composer.className = "agent-chat__composer-combobox";
-    const textarea = composer.appendChild(document.createElement("textarea"));
-    pane.append(composer);
-    const focus = vi.spyOn(textarea, "focus");
-    const button = document.body.appendChild(document.createElement("button"));
-    button.addEventListener("keydown", (event) => pane.handleDocumentKeydown(event));
-    button.focus();
-
-    try {
-      button.dispatchEvent(
-        new KeyboardEvent("keydown", { key: "x", bubbles: true, composed: true }),
-      );
-
-      expect(focus).toHaveBeenCalledWith({ preventScroll: true });
-    } finally {
-      button.remove();
-    }
-  });
-
-  it("ignores an open dropdown in an inactive retained pane", () => {
-    const { pane } = createTestChatPane({
-      client: createGatewayBrowserClientFixture(),
-      sessions: createSessionCapabilityFixture(),
-    });
-    pane.active = true;
-    pane.presented = true;
-    const composer = document.createElement("div");
-    composer.className = "agent-chat__composer-combobox";
-    const textarea = composer.appendChild(document.createElement("textarea"));
-    pane.append(composer);
-    const focus = vi.spyOn(textarea, "focus");
-    const target = pane.appendChild(document.createElement("main"));
-    target.addEventListener("keydown", (event) => pane.handleDocumentKeydown(event));
-    const retainedPane = document.body.appendChild(document.createElement("div"));
-    retainedPane.inert = true;
-    const retainedDropdown = retainedPane.appendChild(
-      document.createElement("wa-dropdown"),
-    ) as HTMLElement & { open: boolean };
-    retainedDropdown.open = true;
-
-    try {
-      target.dispatchEvent(
-        new KeyboardEvent("keydown", { key: " ", bubbles: true, composed: true }),
-      );
-
-      expect(focus).toHaveBeenCalledWith({ preventScroll: true });
-    } finally {
-      retainedPane.remove();
-    }
-  });
-
-  it("distinguishes an open disclosure from an open overlay", () => {
-    const { pane } = createTestChatPane({
-      client: createGatewayBrowserClientFixture(),
-      sessions: createSessionCapabilityFixture(),
-    });
-    pane.active = true;
-    pane.presented = true;
-    const composer = document.createElement("div");
-    composer.className = "agent-chat__composer-combobox";
-    const textarea = composer.appendChild(document.createElement("textarea"));
-    pane.append(composer);
-    const focus = vi.spyOn(textarea, "focus");
-    const details = document.body.appendChild(document.createElement("details"));
-    details.open = true;
-    const summary = details.appendChild(document.createElement("summary"));
-    summary.addEventListener("keydown", (event) => pane.handleDocumentKeydown(event));
-    let dialog: HTMLDialogElement | null = null;
-
-    try {
-      summary.focus();
-      summary.dispatchEvent(
-        new KeyboardEvent("keydown", { key: "x", bubbles: true, composed: true }),
-      );
-      expect(focus).toHaveBeenCalledWith({ preventScroll: true });
-
-      focus.mockClear();
-      dialog = document.body.appendChild(document.createElement("dialog"));
-      dialog.open = true;
-      const dialogButton = dialog.appendChild(document.createElement("button"));
-      dialogButton.addEventListener("keydown", (event) => pane.handleDocumentKeydown(event));
-      dialogButton.focus();
-      dialogButton.dispatchEvent(
-        new KeyboardEvent("keydown", { key: "x", bubbles: true, composed: true }),
-      );
-      expect(focus).not.toHaveBeenCalled();
-    } finally {
-      details.remove();
-      dialog?.remove();
-    }
-  });
-
-  it("does not steal typing focus from a shadow-root confirmation", async () => {
-    const restoreDialogPolyfill = installDialogPolyfill();
-    const { pane } = createTestChatPane({
-      client: createGatewayBrowserClientFixture(),
-      sessions: createSessionCapabilityFixture(),
-    });
-    pane.active = true;
-    pane.presented = true;
-    const composer = document.createElement("div");
-    composer.className = "agent-chat__composer-combobox";
-    const textarea = composer.appendChild(document.createElement("textarea"));
-    pane.append(composer);
-    const focus = vi.spyOn(textarea, "focus");
-    const container = document.body.appendChild(document.createElement("div"));
-    const modal = container.appendChild(document.createElement("openclaw-modal-dialog"));
-    const cancel = modal.appendChild(document.createElement("button"));
-
-    try {
-      const { dialog } = await getRenderedModalDialog(container);
-      expect(dialog.open).toBe(true);
-      expect(document.querySelector("dialog[open]")).toBeNull();
-      cancel.addEventListener("keydown", (event) => pane.handleDocumentKeydown(event));
-
-      cancel.dispatchEvent(new KeyboardEvent("keydown", { key: "x", cancelable: true }));
-
-      expect(focus).not.toHaveBeenCalled();
-    } finally {
-      container.remove();
-      restoreDialogPolyfill();
-    }
-  });
-
-  it("does not steal typing focus from a light-DOM confirmation", () => {
-    const { pane } = createTestChatPane({
-      client: createGatewayBrowserClientFixture(),
-      sessions: createSessionCapabilityFixture(),
-    });
-    pane.active = true;
-    pane.presented = true;
-    const composer = document.createElement("div");
-    composer.className = "agent-chat__composer-combobox";
-    const textarea = composer.appendChild(document.createElement("textarea"));
-    pane.append(composer);
-    const focus = vi.spyOn(textarea, "focus");
-    const modal = document.body.appendChild(document.createElement("div"));
-    modal.setAttribute("aria-modal", "true");
-
-    try {
-      pane.handleDocumentKeydown(new KeyboardEvent("keydown", { key: "x", cancelable: true }));
-
-      expect(focus).not.toHaveBeenCalled();
-    } finally {
-      modal.remove();
-    }
-  });
-
   it("toggles only the active pane's session workspace", () => {
     const client = createGatewayBrowserClientFixture();
     const sessions = createSessionCapabilityFixture();
